Extracted logic from FormFactory to FormWrapper
