+++ src/main/java/dk/eamv/ferrari/sharedcomponents/filter/forms/CRUDType.java
@@ -1,0 +1,7 @@
+package dk.eamv.ferrari.sharedcomponents.filter.forms;
+
+public enum CRUDType {
+    CAR,
+    LOAN,
+    CUSTOMER
+}
